First pass on figuring out how we will demux a signal to many apps
Overview

When a given events happens in lessonly and we notify Demux about it, at
that point Demux will need to find any apps that are installed on the
"account" the event happened in and send a webhook/ signal to any of
those apps that are listening for that event.

- Demux is notified something happened
- Demux finds apps installed on given account (company in our case)
- Demux finds which of these apps are listening for this event
- Demux will now send this webhook/signal out to the apps

Currently, the thought is that the means by which the signals are sent
would be determined by the "demuxer" you have configured. By default, it
would use a base "in memory" demuxer but in production we would likely
use a "sidekiq" based demuxer. It would call the base behavior but
inside of asynchronous jobs.

Where it landed

I think we have a viable API here for defining and sending webhook
"Signals". It would make adding and configuring a new signal super clear
and allow grouping of signals under a name.

We have a basic inline demuxer. This would likely not be used in
production, but the goal was to make it simple to override and add app
specific behavior. In our case, we'll define a demuxer that splits out
into sidekiq jobs.

You can now open a console and call the LessonSignal in the dummy app
and see it work end to end 🎉 We also have one integration spec that
tests it end to end by doing just that and mocking the http endpoints.

Since the API is starting to stabilize and this works in the end to end
happy path, I'm going to stop here and merge before continuing.

Things yet to consider in future work

- Test individual methods and look for edge cases - Performance testing.
It appears from sql explains that looking up by uniqueness hash is more
performant that using a where across several fields; does that continue
to hold up?  - What about signals that need to capture data in the
moment? For example... when a record is destroyed we can't get it's
"title" attribute later because it's gone. How can we pass in the moment
context?  - Add missing inline documentation - Add limits... 10 second
request timeout for example - Capture and record any headers from the
response if they are given - Make user agent and the names of demux
specific headers configurable to something besides demux - What if there
is an internal exception when trying to send signal. Should we capture
that status somehow on the transmission? The request didn't fail because
the system failed before it was sent... but maybe we just want that to
bubble up as it does now?