Build demux apps and connections to facilitate launching to entry url
In demux, apps will contain an entry_url that users can be directed to
for configuring that specific application. It will be an entry point
from the host app to the peripheral app.

The goal of this PR is to build a way for this configuration to be
represented in the Demux app and connection records to power the
existing "configure" button that we have already in app. This will allow
Demux to return an entry_url signed with a JWT using the secret
configured on that app.