Remove Gemfile.lock from repository
We should not be tied to a particular configuration of dependencies when
working on the engine, rather we should be testing based on the limits
set in the gemspec. Gemfile.lock is handy for a project where you are
locking in dependencies, but we do not want to do this since we are
working on a library that will be installed in projects. We want to make
sure that the latest versions of it's dependencies based on the gemspec
resolve and work appropriately.