Make Demuxer easier to override
One of the goals of the demuxer is to be easy to override with your own
implementation (mostly likely so you can process in a job).

To help with that, this separates out `resolve` from `resolve_now` so
that you can override `resolve` and still use `resolve_now` in your
implementation.

It also separates out `transmit` into a method so that you can override
this as well.

I also went with the name `resolve` over the former `send_to_apps`
because eventually we might be sending to more than apps and that name
was too restrictive. What if someday, we also resolve to individual
webhooks in addition to apps?