Add a configurable timeout for sending signals
closes #7

Client apps should process signals sent to them in a timely fashion
because it consumes resources from the host application and this is made
much worse if there is dead time while they client app is processing.

In general, client apps should receive a signal in a non blocking way
and return a response as soon as possible.

To help encourage this behavior, the default timeout is 10 seconds.

This timeout period is configurable by the host app.