Refactor reference signal_attributes to not share instance variable
When subclasses access an instance variable from a parent, it makes
assumptions about that instance variable. This couples the two together
in a way that isn't clear to understand and can make change had.

This refactor adds a method that can be called instead of accessing the
instance variable.

Additionally, it's a little awkward to have to know about the
SignalAttributes object in the custom demuxer. This changes the demuxer
to initialize using keyword arguments instead of a SignalAttributes
object.