Accept extra data to include in token when building entry_url
Resolves #8

When building a signed entry url from a connection, it could be helpful
to pass additional data to be sent in the token. For example, maybe in
addition to an account_id you would want to pass the user_id of the user
that launched the app.

It's worth noting that you can already do this by using the
`signed_entry_url` method defined on an instance of an app and supply
any data you want there. This is specifically for cases where you want
the entry url from a connection (including the account_id automatically)
but would also like to pass additional data.